backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
pydantic-settings==2.6.1
stripe==11.0.0
python-multipart==0.0.12
pytest==8.3.3
httpx==0.27.0
requests==2.32.3

